+++ js/displaying-thumbnails.js
@@ -4,12 +4,10 @@ const childElements = document.querySelector('.pictures');
 const createThumbnail = ( {id, url, description, likes, comments} ) => {
   const thumbnail = patternThumbnails.cloneNode(true);
 
-
   thumbnail.querySelector('.picture__img').src = url;
   thumbnail.querySelector('.picture__img').alt = description;
   thumbnail.querySelector('.picture__likes').textContent = likes;
   thumbnail.querySelector('.picture__comments').textContent = comments.length;
-  thumbnail.dataset.thumbnailId = id;
   thumbnail.dataset.thumbnailId = id;
 
   return thumbnail;
