+++ js/main.js
@@ -1,6 +1,6 @@
 import { getPictures } from './data.js';
 import { renderGallery } from './rendering-full-size-image.js';
-import './form-data.js';
+import './validation.js';
 
 renderGallery(getPictures());
